IntroToPython/aPyThing_TeleBot/requirements.txt
python-telegram-bot==13.1
requests==2.25.1

